--- src/main/java/com/company/cpluss/Configurations/securityConfig/AuthenticationRequest.java
@@ -1,13 +1,0 @@
-package com.company.cpluss.Configurations.securityConfig;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthenticationRequest {
-    String username;
-    String password;
-}
