--- src/main/java/com/company/cpluss/Configurations/securityConfig/AuthenticationResponse.java
@@ -1,12 +1,0 @@
-package com.company.cpluss.Configurations.securityConfig;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationResponse {
-    private String jwt;
-}
