--- src/main/java/com/company/cpluss/Configurations/securityConfig/MyUserDetailsService.java
@@ -1,64 +1,0 @@
-package com.company.cpluss.Configurations.securityConfig;
-
-import com.company.cpluss.repository.UserRepository;
-import com.company.cpluss.model.userInfo.Role;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
-
-@Service
-public class MyUserDetailsService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    public MyUserDetailsService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public User loadUserByUsername(String userName) throws UsernameNotFoundException {
-
-        Optional<com.company.cpluss.model.userInfo.User> authuser = userRepository.findByUsername(userName);
-        authuser.orElseThrow(()-> new UsernameNotFoundException("user does not exist"));
-
-        Optional<com.company.cpluss.model.userInfo.User> authuser1 = userRepository.findByEmail(userName);
-        authuser.orElseThrow(()-> new UsernameNotFoundException("user does not exist"));
-
-        List<GrantedAuthority> authorities;
-
-        if(authuser.isPresent() ) {
-            authorities = authuser
-                    .get().getRoles().stream()
-                    .map(Role::getRoleName)
-                    .map(SimpleGrantedAuthority::new)
-                    .collect(Collectors.toList());
-            //signin using username
-
-            return new User(authuser.get().getUsername(), authuser.get().getPassword(), authorities);
-        }else {
-            authorities = authuser1
-                    .get().getRoles().stream()
-                    .map(Role::getRoleName)
-                    .map(SimpleGrantedAuthority::new)
-                    .collect(Collectors.toList());
-            //signin using email
-
-            return new User(authuser1.get().getUsername(), authuser1.get().getPassword(), authorities);
-        }
-
-
-    }
-
-
-
-}
