+++ src/main/java/com/company/cpluss/CplussApplication.java
@@ -2,6 +2,7 @@ package com.company.cpluss;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 
 @SpringBootApplication
 public class CplussApplication {
