--- src/main/java/com/company/cpluss/controller/OrderController.java
@@ -1,18 +1,0 @@
-package com.company.cpluss.controller;
-
-import com.company.cpluss.service.OrderService;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/orders")
-public class OrderController {
-
-    private final OrderService orderService;
-
-    public OrderController(OrderService orderService) {
-        this.orderService = orderService;
-    }
-
-
-}
