--- src/main/java/com/company/cpluss/controller/ProductController.java
@@ -1,28 +1,0 @@
-package com.company.cpluss.controller;
-
-import com.company.cpluss.model.items.Product;
-import com.company.cpluss.service.ProductService;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/products")
-public class ProductController {
-
-    private final ProductService productService;
-
-    public ProductController(ProductService productService) {
-        this.productService = productService;
-    }
-
-
-    @PostMapping
-    public ResponseEntity addProduct(@RequestBody Product product){
-        return ResponseEntity.ok(productService.addProduct(product));
-    }
-
-
-}
