+++ src/main/java/com/company/cpluss/inventory_service/model/Product.java
@@ -1,0 +1,30 @@
+package com.company.cpluss.inventory_service.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Data
+@Entity
+public class Product {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    private String productID;
+    private String productName;
+    private double amount;
+    private String productDescription;
+    private String sellerName;
+
+
+
+}
