+++ src/main/java/com/company/cpluss/inventory_service/repository/AddressRepo.java
@@ -1,0 +1,7 @@
+package com.company.cpluss.inventory_service.repository;
+
+import com.company.cpluss.security.model.Address;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AddressRepo extends JpaRepository<Address, Long> {
+}
