+++ src/main/java/com/company/cpluss/inventory_service/repository/OrderRepository.java
@@ -1,0 +1,8 @@
+package com.company.cpluss.inventory_service.repository;
+
+import com.company.cpluss.inventory_service.model.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+    Order findByOrderID(String orderID);
+}
