--- src/main/java/com/company/cpluss/model/items/Order.java
@@ -1,30 +1,0 @@
-package com.company.cpluss.model.items;
-
-import com.sun.istack.Nullable;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToOne;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@Entity
-public class Order {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    private String orderID ;
-    @OneToOne
-    private Product product;
-    private float quantity;
-    @Nullable
-    private double totalAmount;// = quantity* product.getAmount();
-
-}
