--- src/main/java/com/company/cpluss/model/userInfo/Address.java
@@ -1,27 +1,0 @@
-package com.company.cpluss.model.userInfo;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@SuperBuilder
-public class Address {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    private String district;
-    private String city;
-    private String village;
-}
