--- src/main/java/com/company/cpluss/model/userInfo/Role.java
@@ -1,24 +1,0 @@
-package com.company.cpluss.model.userInfo;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@Data
-@SuperBuilder
-public class Role {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    private String roleName;
-}
