+++ src/main/java/com/company/cpluss/notification_service/model/Email.java
@@ -1,0 +1,16 @@
+package com.company.cpluss.notification_service.model;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Email {
+
+    private String to;
+    private String subject;
+    private String body;
+
+}
