--- src/main/java/com/company/cpluss/repository/OrderRepository.java
@@ -1,7 +1,0 @@
-package com.company.cpluss.repository;
-
-import com.company.cpluss.model.items.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface OrderRepository extends JpaRepository<Order, Long> {
-}
