--- src/main/java/com/company/cpluss/repository/UserRepository.java
@@ -1,11 +1,0 @@
-package com.company.cpluss.repository;
-
-import com.company.cpluss.model.userInfo.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUsername(String userName);
-    Optional<User> findByEmail(String email);
-}
