+++ src/main/java/com/company/cpluss/security/repository/RoleRepository.java
@@ -1,0 +1,7 @@
+package com.company.cpluss.security.repository;
+
+import com.company.cpluss.security.model.Role;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface RoleRepository extends JpaRepository<Role, Long> {
+}
