+++ src/main/java/com/company/cpluss/security/repository/UserRepository.java
@@ -1,0 +1,8 @@
+package com.company.cpluss.security.repository;
+
+import com.company.cpluss.security.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+    User findByUsername(String userName);
+}
