+++ src/main/java/com/company/cpluss/security/securityConfig/AuthenticationRequest.java
@@ -1,0 +1,13 @@
+package com.company.cpluss.security.securityConfig;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class AuthenticationRequest {
+    String username;
+    String password;
+}
