+++ src/main/java/com/company/cpluss/security/securityConfig/AuthenticationResponse.java
@@ -1,0 +1,12 @@
+package com.company.cpluss.security.securityConfig;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class AuthenticationResponse {
+    private String jwt;
+}
