+++ src/main/java/com/company/cpluss/security/securityConfig/SecurityConfigurer.java
@@ -1,0 +1,65 @@
+package com.company.cpluss.security.securityConfig;
+
+import com.company.cpluss.security.securityConfig.jwt.JWTRequestFilter;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+@Configuration
+@EnableWebSecurity
+public class SecurityConfigurer extends WebSecurityConfigurerAdapter {
+    @Autowired
+    private MyUserDetailsService myUserDetailsService;
+
+    @Autowired
+    private JWTRequestFilter jwtRequestFilter;
+
+    public SecurityConfigurer(MyUserDetailsService myUserDetailsService) {
+        this.myUserDetailsService = myUserDetailsService;
+    }
+
+    @Override
+    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+        auth.userDetailsService(myUserDetailsService);
+    }
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception{
+        http.csrf().disable()
+                .authorizeRequests().antMatchers("/users/auth").permitAll()
+                .antMatchers("/users/register").permitAll()
+                //.antMatchers("/c1/users").permitAll()  //works if you also want to except this one from the auth section
+                .anyRequest().authenticated()
+        .and().sessionManagement()
+        .sessionCreationPolicy(SessionCreationPolicy.STATELESS);
+        http.addFilterBefore(jwtRequestFilter, UsernamePasswordAuthenticationFilter.class);
+    }
+
+    @Bean
+    @Override
+    public AuthenticationManager authenticationManagerBean() throws Exception {
+        return super.authenticationManagerBean();
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder(){
+        //return NoOpPasswordEncoder.getInstance();
+        return new BCryptPasswordEncoder(12);
+    }
+
+//    @Bean
+//    GrantedAuthorityDefaults grantedAuthorityDefaults(){
+//        return new GrantedAuthorityDefaults("");
+//    }
+
+
+}
