--- src/main/java/com/company/cpluss/service/OrderService.java
@@ -1,17 +1,0 @@
-package com.company.cpluss.service;
-
-import com.company.cpluss.repository.OrderRepository;
-import org.springframework.stereotype.Service;
-
-@Service
-public class OrderService {
-
-    private final OrderRepository orderRepository;
-
-    public OrderService(OrderRepository orderRepository) {
-        this.orderRepository = orderRepository;
-    }
-
-
-
-}
