--- src/main/java/com/company/cpluss/service/ProductService.java
@@ -1,19 +1,0 @@
-package com.company.cpluss.service;
-
-import com.company.cpluss.model.items.Product;
-import com.company.cpluss.repository.ProductRepository;
-import org.springframework.stereotype.Service;
-
-@Service
-public class ProductService {
-
-    private final ProductRepository productRepository;
-
-    public ProductService(ProductRepository productRepository) {
-        this.productRepository = productRepository;
-    }
-
-    public Product addProduct(Product product){
-        return productRepository.save(product);
-    }
-}
