+++ src/main/resources/application.properties
@@ -1,7 +1,16 @@
-server.port = 9000
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/cpluss
+server.port = 8899
+
+spring.datasource.url=jdbc:mysql://localhost:3306/cpluss
 spring.datasource.username=root
 spring.datasource.password=APLEXY
-spring.datasource.driver-class-name =com.mysql.cj.jdbc.Driver
-spring.jpa.show-sql = true
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+spring.jpa.database-platform = org.hibernate.dialect.MySQL5Dialect
+#spring.jpa.generate-ddl=true
+spring.jpa.hibernate.ddl-auto = update
+
+spring.mail.host=smtp.gmail.com
+spring.mail.port=587
+spring.mail.username=cpluss.delivery@gmail.com
+spring.mail.password=kngrqbzjxqrgsfyq
+spring.mail.properties.mail.smtp.auth=true
+spring.mail.properties.mail.smtp.starttls.enable=true
